pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "laneye"
version = "1.1.0"
description = "Lightweight Network IP Scanner with Web GUI"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "SiteQ8", email = "site@hotmail.com"},
]
keywords = ["network", "scanner", "lan", "ip", "arp", "monitoring", "security"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Networking :: Monitoring",
    "Topic :: Security",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "aiosqlite>=0.20.0",
    "scapy>=2.6.1",
    "aiohttp>=3.10.0",
    "pydantic>=2.9.0",
    "prometheus-client>=0.21.0",
    "python-dotenv>=1.0.1",
    "PyYAML>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.5",
]

[project.urls]
Homepage = "https://github.com/SiteQ8/LANEye"
Documentation = "https://siteq8.github.io/LANEye/"
Repository = "https://github.com/SiteQ8/LANEye"
Issues = "https://github.com/SiteQ8/LANEye/issues"

[project.scripts]
laneye = "main:app"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
target-version = "py39"
line-length = 100
